Big changes!
- Location is now a 3D coordinate.
- GameObject has been introduced, 'blocks' of the map are now
  GameObjects, as well as anything else that will be added that
  has a position and size.
- GraphicsComponent has been introduced to handle how an GameObject
  appears. Have only implemented StaticGraphicsComponent.

This means that things still need to be moved around to fit properly,
like who and where creates the GraphicsComponents? Also the rendering
doesn't really make sense... Location is based on map tile positions
and the draw coordinate is calulated using the tile width and height
specified. But this means that location can be linked to a tile,
something can't really move across a tile...