Added Weather
Add rain to the terrain. We currently add clouds in the south and these
move northwards and begin to dissipate once they hit land. The clouds
split into two once they encounter a terrace higher than the one
currently occupied.

The biome calculation is now a little more complex, with three levels
of wetness and two heights for land.