Improved depth perception
Added 'Edge' to TerrainShape and sprites with have a line of shadow
on them to highlight the edges to stop everything from looking so
flat. I've rewritten how we insert the ramps, which is looking quite
good, except that there doesn't appear to be any ramps coming from
the east.