MoveAction
Added a class to perform movement. It will need to receive a
destination which is waypoint from the current position that only
changes on, at maximum, two axes at one time. This limitation makes
it easier for my brain and just means that movement up and down ramps
can only also on the x or y plane.